+++ lib/src/generators/router/generator/arguments_class/arguments_class_builder.dart
@@ -35,8 +35,45 @@ class ArgumentsClassBuilder {
                   (b) => b..symbol = 'String',
                 ),
             ),
+            _buildEqualityOperator(
+              className: argumentsBuilderHelper.argumentClassName,
+              fields: argumentsBuilderHelper.convertParametersToClassFields,
+            ),
+            _buildHashCode(
+                argumentsBuilderHelper.convertParametersToClassFields),
           ]),
       );
     });
   }
+}
+
+Method _buildHashCode(List<Field> fields) {
+  final buffer = StringBuffer();
+  buffer.writeln(
+      'return ${fields.map((f) => '${f.name}.hashCode').join(' ^ ')};');
+
+  return Method((b) => b
+    ..name = 'hashCode'
+    ..returns = refer('int')
+    ..annotations.add(refer('override'))
+    ..type = MethodType.getter
+    ..body = Code(buffer.toString().trim()));
+}
+
+Method _buildEqualityOperator(
+    {required String className, required List<Field> fields}) {
+  final buffer = StringBuffer();
+  buffer.writeln('if (identical(this, other)) return $literalTrue;');
+  buffer.writeln(
+      'return ${fields.map((f) => 'other.${f.name} == ${f.name}').join(' && ')};');
+
+  return Method((b) => b
+    ..name = 'operator =='
+    ..returns = refer('bool')
+    ..annotations.add(refer('override'))
+    ..requiredParameters.add(Parameter((p) => p
+      ..covariant = true
+      ..name = 'other'
+      ..type = refer(className)))
+    ..body = Code(buffer.toString()));
 }