+++ lib/src/generators/router_2/router_extension_builder/router_extension_builder_helper.dart
@@ -3,9 +3,8 @@ import 'package:stacked_generator/src/generators/extensions/string_utils_extensi
 import 'package:stacked_generator/src/generators/router_2/code_builder/route_info_builder.dart';
 import 'package:stacked_generator/src/generators/router_common/models/route_config.dart';
 import 'package:stacked_generator/src/generators/router_common/models/route_parameter_config.dart';
-import 'package:stacked_generator/utils.dart';
 
-mixin NavigateExtensionClassBuilderHelper {
+mixin RouterExtensionBuilderHelper {
   Iterable<Method> buildNavigateToExtensionMethods(
     List<RouteConfig> routes,
     DartEmitter emitter,
@@ -19,7 +18,9 @@ mixin NavigateExtensionClassBuilderHelper {
   }
 
   Method extractNavigateToMethodFromRoute(
-          RouteConfig route, DartEmitter emitter) =>
+    RouteConfig route,
+    DartEmitter emitter,
+  ) =>
       _extractNavigationMethod(
         navigationMethod: 'navigateTo',
         route: route,
@@ -27,7 +28,9 @@ mixin NavigateExtensionClassBuilderHelper {
       );
 
   Method extractReplaceWithMethodFromRoute(
-          RouteConfig route, DartEmitter emitter) =>
+    RouteConfig route,
+    DartEmitter emitter,
+  ) =>
       _extractNavigationMethod(
         navigationMethod: 'replaceWith',
         route: route,
@@ -39,23 +42,23 @@ mixin NavigateExtensionClassBuilderHelper {
     required RouteConfig route,
     required DartEmitter emitter,
   }) {
-    final methodName = route.parentClassName != null
-        ? '${navigationMethod}Nested${route.name?.capitalize}In${route.parentClassName}'
-        : '$navigationMethod${route.name?.capitalize}';
+    final methodName = '$navigationMethod${route.name?.capitalize}';
 
     final methodReturnType = route.isProcessedReturnTypeDynamic
         ? route.processedReturnType
         : '${route.processedReturnType}?';
 
-    final viewArgumentsParameter = notQueryNorPath(route.parameters)
-        .map((parameter) => _extractViewArgumentsParametrs(parameter, emitter));
+    final viewArgumentsParameter = route.parameters.map(
+        (parameter) => _extractViewArgumentsParameters(parameter, emitter));
 
     return Method((b) => b
       ..name = methodName
       ..modifier = MethodModifier.async
       ..returns = Reference('Future<$methodReturnType>')
-      ..optionalParameters
-          .addAll([...viewArgumentsParameter, ..._constParameters])
+      ..optionalParameters.addAll([
+        ...viewArgumentsParameter,
+        ..._constParameters,
+      ])
       ..body = _body(
         route: route,
         methodReturnType: methodReturnType,
@@ -64,8 +67,10 @@ mixin NavigateExtensionClassBuilderHelper {
   }
 
   /// The arguments provided to the view
-  Parameter _extractViewArgumentsParametrs(
-      ParamConfig param, DartEmitter emitter) {
+  Parameter _extractViewArgumentsParameters(
+    ParamConfig param,
+    DartEmitter emitter,
+  ) {
     return Parameter((parameterBuilder) {
       parameterBuilder
         ..name = param.name
@@ -88,27 +93,21 @@ mixin NavigateExtensionClassBuilderHelper {
 
   /// These are the parameters that exists on every call
   ///
-  /// int? routerId,
-  /// bool preventDuplicates = true,
-  /// Map<String, String>? parameters,
-  /// Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)? transition,
+  /// void Function(NavigationFailure)? onFailure,
   List<Parameter> get _constParameters => [
-        Parameter((b) => b
-          ..name = 'routerId'
-          ..type = const Reference('int?')),
-        Parameter((b) => b
-          ..name = 'preventDuplicates'
-          ..type = const Reference('bool')
-          ..defaultTo = const Code('true')),
-        Parameter((b) => b
-          ..name = 'parameters'
-          ..type = const Reference('Map<String, String>?')),
         Parameter(
           (b) => b
-            ..name = 'transition'
-            ..type = const Reference(
-                'Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?'),
-        )
+            ..name = 'onFailure'
+            ..named = true
+            ..type = FunctionType((b) => b
+              ..symbol = 'Function'
+              ..requiredParameters.add(refer(
+                'NavigationFailure',
+                'package:stacked/stacked.dart',
+              ))
+              ..isNullable = true
+              ..returnType = refer('void')),
+        ),
       ];
 
   Code _body({
@@ -116,34 +115,27 @@ mixin NavigateExtensionClassBuilderHelper {
     required String methodReturnType,
     required String navigationMethod,
   }) {
-    final routesClassName = route.parentRouterConfig != null
-        ? route.parentRouterConfig!.routesClassName
-        : 'Routes';
+    var appendOrNotConst = 'const ';
+    var appendOrNotParameters = '';
+
+    if (route.parameters.isNotEmpty) {
+      appendOrNotConst = '';
+      appendOrNotParameters = route.parameters
+          .map((p) => '${p.name}: ${p.name},')
+          .join('')
+          .toString();
+    }
+
     return Block.of([
       Code(
-          'return $navigationMethod<$methodReturnType>($routesClassName.${route.name}, '),
-      _assignArgumentClass(route),
+        'return $navigationMethod($appendOrNotConst${route.className}Route($appendOrNotParameters),',
+      ),
       ..._constMethodBodyParameters,
     ]);
   }
 
   List<Code> get _constMethodBodyParameters => [
-        const Code('id: routerId,'),
-        const Code('preventDuplicates: preventDuplicates,'),
-        const Code('parameters: parameters,'),
-        const Code('transition: transition'),
+        const Code('onFailure: onFailure,'),
         const Code(');'),
       ];
-
-  Code _assignArgumentClass(RouteConfig route) {
-    if (route.parameters.isNotEmpty) {
-      final argumentClassName = route.parentClassName != null
-          ? 'Nested${route.argumentsHolderClassName}'
-          : route.argumentsHolderClassName;
-
-      return Code(
-          'arguments: $argumentClassName(${route.parameters.map((p) => '${p.name}: ${p.name}').join(',')}),');
-    }
-    return const Code('');
-  }
 }